Remove redunancies between square & rect buttons
